chore: migrate zod validation to dedicated package
