+++ packages/form-core/src/index.ts
@@ -1,3 +1,4 @@
 export * from './FormApi'
 export * from './FieldApi'
 export * from './utils'
+export * from './types'
