+++ packages/form-zod-adapter/.eslintrc.cjs
@@ -1,0 +1,11 @@
+// @ts-check
+
+/** @type {import('eslint').Linter.Config} */
+const config = {
+  parserOptions: {
+    tsconfigRootDir: __dirname,
+    project: './tsconfig.eslint.json',
+  },
+}
+
+module.exports = config
