+++ packages/form-zod-adapter/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@tanstack/form-core",
+  "name": "@tanstack/form-zod-adapter",
   "version": "0.3.6",
-  "description": "Powerful, type-safe, framework agnostic forms.",
+  "description": "The Zod adapter for TanStack Form.",
   "author": "tannerlinsley",
   "license": "MIT",
   "repository": "tanstack/form",
@@ -51,6 +51,12 @@
     "build": "tsup"
   },
   "dependencies": {
-    "@tanstack/store": "0.1.3"
+    "@tanstack/form-core": "workspace:*"
+  },
+  "peerDependencies": {
+    "zod": "^3.x"
+  },
+  "devDependencies": {
+    "zod": "^3.22.4"
   }
 }
