+++ packages/form-zod-adapter/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "composite": true,
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "outDir": "./build/lib",
+    "types": ["vitest/globals"]
+  },
+  "include": ["src"]
+}
